# Remove @dataclass, add slots