--- a/.gitignore
+++ b/.gitignore
@@ -16,3 +16,4 @@ output
 .tox
 .venv
 .asv
+venv