--- a/src/betterproto/__init__.py
+++ b/src/betterproto/__init__.py
@@ -540,8 +540,11 @@ class MessageMeta(ABCMeta):
         setattr(
             message_class, dataclasses._FIELDS, fields,
         )
+        setattr(
+            message_class, dataclasses._PARAMS, fields,
+        )
         # we don't need to set __dataclass_params__ as its only use appears to be for
-        # checking if any base clasess are frozen
+        # checking if any base classes are frozen
         return message_class
 
 
@@ -648,6 +651,7 @@ class Message(metaclass=MessageMeta):
         return value
 
     def __setattr__(self, attr: str, value: Any) -> None:
+        print("setting", attr)
         if attr != "_serialized_on_wire":
             # Track when a field has been set.
             self._serialized_on_wire = True