--- a/src/betterproto/__init__.py
+++ b/src/betterproto/__init__.py
@@ -651,7 +651,6 @@ class Message(metaclass=MessageMeta):
         return value
 
     def __setattr__(self, attr: str, value: Any) -> None:
-        print("setting", attr)
         if attr != "_serialized_on_wire":
             # Track when a field has been set.
             self._serialized_on_wire = True