… .github/CONTRIBUTING.md
… .github/workflows/code-quality.yml
… .readthedocs.yml
… README.md
… asv.conf.json
… benchmarks/__init__.py
… benchmarks/benchmarks.py
… docs/api.rst
… docs/index.rst
… docs/migrating.rst
… docs/quick-start.rst
… docs/upgrading.md
… poetry.lock
… pyproject.toml
… src/TEXT_TEST_PY
… src/betterproto/_types.py
… src/betterproto/casing.py
… src/betterproto/compile/importing.py
… src/betterproto/compile/naming.py
… src/betterproto/grpc/grpclib_client.py
… src/betterproto/grpc/util/async_channel.py
… src/betterproto/lib/google/protobuf/__init__.py
… src/betterproto/plugin/compiler.py
… src/betterproto/plugin/main.py
… src/betterproto/plugin/models.py
… src/betterproto/plugin/parser.py
… src/betterproto/templates/template.py.j2
… src/betterproto/utils.py
… tests/grpc/test_stream_stream.py
… tests/mocks.py