Adds a custom metaclass for Message to quickly create dataclasses for Messages whilst also slotting them. Internally should still be basically the same as a dataclass.

Benefits:

- This will almost certainly be faster than the current approach, I will need to run speed tests on it before I am 100% sure.
- Removes an import from all generated messages, and improves visual clarity.

This is incorporates 2 breaking changes:

- Removing `@dataclass` from all generated outputs, since this is for V2 however I think it is fine, there are a couple of options if you want to give warning like monkey patching the dataclass decorator to do nothing for Message subclasses.
- Fields cannot now be access outside of instances of themselves, pretty minor change, this shouldn't really affect anyone as they (I think) would always be `betterproto.Placeholder`, but if necessary it could be remedied by implementing `MessageMeta.__getattribute__`.

Closes #50