@abn 2020-08-27
I am not sure if optimising for performance over usability by default is a good thing. While, I can understand the motivation for this change, I feel that dropping dataclasses means you are loosing a lot of niceties that it brings to the table. I am also unclear on how much of a gain we are talking here. We might also want to talk about making it optional, ie. dataclasses vs slots at compile time.

As an additional note, the alternative is to move from dataclasses to something that already fascilitates slots. The [attrs](https://www.attrs.org/en/stable/examples.html#slots) project is a good option for that. Obviously, everything has trade-offs. :)