@Gobot1234 2020-08-27
It isn't dropping anything internally, it just removes the decorator.