@abn 2020-08-27
> It isn't dropping anything internally, it just removes the decorator.

Ah, I might have spoke too soon then. Will go through the changes. Ignore my reactionary comment. :)