@Gobot1234 2020-08-27
Some tests

# Results

Tabled results are for a Message of unless specified otherwise:
```py
class Message(betterproto.Message):
    foo: int = betterproto.uint32_field(0)
    bar: int = betterproto.uint32_field(1)
    baz: int = betterproto.uint32_field(2)
```

## Important results:

|  | Old implementation |New slotted implementation |
|--|--|--|
| Size of [1](#size-of-instance) | 576 bytes | 256 bytes |
| Runtime Creation [2](#runtime-overhead) | 741 usec | 38.4 usec |
| Attribute Access [3](#size-of-instance) | 71.4 nsec |  43.8 nsec |


### Size of Instance

Tested with 0, 3, 5, 10, 50 field large message instances.

| Size | 0 | 3 | 5 | 10 | 50 | 
|--|--|--|--|--|--|
| Old | 152 | 576 | 728 | 1096 | 3848 |
| New | 208 | 256 | 272 | 312 | 632 |

Slots make messages significantly more linear in the amount of memory usage they end up using 

### Runtime overhead

```
Old raw times: 74.1 sec, 90.5 sec, 74.1 sec, 82 sec, 87.8 sec

100000 loops, best of 5: 741 usec per loop
```
```
Slotted raw times: 4.96 sec, 4.72 sec, 4.17 sec, 3.84 sec, 4.1 sec

100000 loops, best of 5: 38.4 usec per loop
```
This is ridiculously faster with MessageMeta at around 20x

### Attribute Access

```
Old raw times: 13.2 msec, 8.81 msec, 7.48 msec, 7.28 msec, 7.14 msec

100000 loops, best of 5: 71.4 nsec per loop
```
```
Slotted raw times: 4.38 msec, 6.24 msec, 5.9 msec, 4.46 msec, 5.67 msec

100000 loops, best of 5: 43.8 nsec per loop
```
So it's about 60% faster with slots

### Instantiation 
```
Old raw times: 2.34 sec, 1.03 sec, 1 sec, 1.61 sec, 2.55 sec

100000 loops, best of 5: 10 usec per loop
```
```
Slotted raw times: 1.25 sec, 1.1 sec, 1.14 sec, 1.12 sec, 1.2 sec

100000 loops, best of 5: 11 usec per loop
```
Currently instantiation with the current implementation is faster, however only marginally and I think the instantiation can be improved on the new model, or if not we can just use the dataclasses implementation, if it doesn't hurt runtime overhead particularly.