@Gobot1234 2020-08-29
I've optimized init more and it's now faster (by a couple of usec) than the dataclasses approach.