@adriangb 2020-09-04
Have you evaluated [Pydantic](https://pydantic-docs.helpmanual.io/) instead? I'm not sure how it would work out speed-wise, but making `Message` inherit from Pydantic's `BaseModel` could be interesting.