@Gobot1234 2020-09-04
What benefits would using pydantic bring over this implementation, I really don't see how it could get much better in terms of performance which is what this is aiming for.