@adriangb 2020-09-04
I don't *think* it will be better in terms of performance (although it might be worth trying) but Pydantic does have a very rich feature set that IMO could be very useful.

Just curious, I see that this comes from #50. Has there been any use case studied where the performance was an issue and this level of performance improvement resolved things?