@Gobot1234 2020-09-04
I don't think performance is a particularly large issue but I see no harm in optimising this kind of stuff.