@adriangb 2020-09-04
> I don't think performance at runtime is a particularly large issue but I see no harm in optimising this kind of stuff.

There is a **lot** of harm in implementing features (performance optimization is a feature) for the sake of implementing features. Without concrete real-world use cases for implementing a feature, it should not be implemented.