@Gobot1234 2020-09-04
I don't agree with that in this case as this **does not** change anything internally. In my eyes better performance is better performance. I thing my previous comment might have slightly down played the benefits this adds in terms of attribute access and memory usage.