@adriangb 2020-09-04
> this **does not** change anything internally.

I see several hundred lines of pretty complex code being changed (without counting the removal of `@dataclass`).

> the benefits this adds in terms of attribute access and memory usage.

Do you have a real world use case where that performance difference is actually important? Does this level of performance improvement resolve or enable those use cases? We could get even better performance by re-writing this in C if we wanted to, but we shouldn't do that just because we can.