@adriangb 2020-09-07
Hey @Gobot1234, would you be able to post your benchmarks? I'd be happy to turn them into [asv](https://github.com/airspeed-velocity/asv) benchmarks and submit that as a separate PR so that it's easy for anyone to evaluate the impact their proposed changes have on performance.