@adriangb 2020-09-07
I added these benchmarks into #148