@adriangb 2020-09-07
Another option to consider: replace the `dataclasses` dependency with [dataclassy](https://github.com/biqqles/dataclassy) which already has the ability to use `__slots__` to speed things up _and_ is compatible with Python 3.6 _and_ has other nice features.