@adriangb 2020-09-13
> We could get even better performance by re-writing this in C if we wanted to

I tested the output from the built in protoc python plugin, and it is an order of magnitude faster than betterproto (because it's ugly, inflexible and has C/C++ portions).