@Gobot1234 2020-09-13
Yes I would have assumed so, that's why I wanted to write an accelerator in Rust