@Gobot1234 2020-09-14
There are also still places I can see room for optimisations, however with how polarising this PR has been I wouldn't expect them to be particularly likely to be merged.