@Gobot1234 2020-10-19
So when I said fields can only be accessed outside of instances of themselves the fields when not set are `types.MemberDescriptor` so field info cannot be accessed.