@Gobot1234 2020-11-07
poe bench returns 
```
All benchmarks:

       before           after         ratio
     [f10bec47]       [222a105b]
     <master>         <dataslots>
      1.99±0.09μs      1.93±0.2μs      0.96 benchmarks.BenchMessage.time_attribute_access
      10.5±0.4μs       9.01±0.2μs      0.86 benchmarks.BenchMessage.time_attribute_setting
      17.3±1μs         13.1±0.2μs      0.76 benchmarks.BenchMessage.time_init_with_values
      17.4±0.5μs       7.96±0.6μs      0.46 benchmarks.BenchMessage.time_instantiation
      722±30μs          293±9μs        0.41 benchmarks.BenchMessage.time_overhead
      22.5±2μs          22.2±1μs       0.99 benchmarks.BenchMessage.time_serialize
	    928               328          0.35 benchmarks.MemSuite.mem_instance
```