@Gobot1234 2020-11-07
Not too sure why the getattr call isn't much of an improvement as it should be the same just `__slot__` boosted?