@Gobot1234 2021-01-03
I'm going to close this for now. I'm sure this can be done better but it requires a lot of work.